working slurm script to process data on Beluga

+++ code/process_connectomes.py
@@ -5,8 +5,9 @@ import json
 from utils.dataset import fetch_fmriprep_derivative, deconfound_connectome_single_strategy, ds000288_movement
 from utils.atlas import create_atlas_masker
 
+
 # define path of input and output
-STRATEGY = "./code/benchmark_strategies.json"
+STRATEGY = "{home}/projects/rrg-pbellec/hwang1/fmriprep-denoise-benchmark/code/benchmark_strategies.json"
 INPUT_FMRIPREP = "{home}/scratch/test_data/1637790137/fmriprep"
 INPUT_BIDS_PARTICIPANTS = "{home}/projects/rrg-pbellec/hwang1/test_data/participants.tsv"
 ATLAS = 'difumo'
@@ -19,7 +20,7 @@ def parse_args():
         description="Generate connectome based on denoising strategy for ds000288.",
     )
     parser.add_argument(
-        "output-path",
+        "output_path",
         action="store",
         type=str,
         help="output path for connectome."
@@ -46,17 +47,18 @@ def parse_args():
 
 def main():
     args = parse_args()
-    print("\n### Running fmriprep-slurm\n")
     print(vars(args))
     strategy_name = args.strategy_name
     atlas_name = args.atlas
-    nroi = args.nroi
-    fmriprep_specifier = "task-pixar"
+    nroi = args.dimension
 
-    strategy_file = Path(__file__).parents[1] / STRATEGY
+    fmriprep_specifier = "task-pixar"
+    
     home = str(Path.home())
+    strategy_file = Path(STRATEGY.format_map({'home': home}))
     input_fmriprep = Path(INPUT_FMRIPREP.format_map({'home': home}))
     input_bids_participants = Path(INPUT_BIDS_PARTICIPANTS.format_map({'home': home}))
+    
     output = Path(args.output_path)
     output.mkdir(exist_ok=True)
 
@@ -65,23 +67,30 @@ def main():
 
     data_aroma = fetch_fmriprep_derivative(input_bids_participants, input_fmriprep,
                                            fmriprep_specifier, aroma=True)
+    if not Path(output / "dataset-ds000288_desc-movement_phenotype.tsv").is_file():
+        movement = ds000288_movement(data)
+        movement.to_csv( output / "dataset-ds000288_desc-movement_phenotype.tsv", sep='\t')
+
     # read the strategy deining files
     with open(strategy_file, "r") as file:
         benchmark_strategies = json.load(file)
-
-    movement = ds000288_movement(data)
-    movement.to_csv( output / "dataset-ds000288_desc-movement_phenotype.tsv", sep='\t')
+    if strategy_name is None:
+        print("Process all strategies.")
+        strategy_names = [*benchmark_strategies]
+    else:
+        strategy_names = [strategy_name]
 
     atlas = create_atlas_masker(atlas_name)
     resolutions = atlas['resolutions'] if nroi is None else [nroi]
     for nroi in resolutions:
         print(f"-- {atlas_name}: dimension {nroi} --")
-        parameters = benchmark_strategies[strategy_name]
-        print(f"Denoising: {strategy_name}")
-        strategy = {strategy_name: parameters}
-        func_data = data_aroma.func if "aroma" in strategy_name else data.func
-        dataset_connectomes = deconfound_connectome_single_strategy(func_data, atlas[nroi]['masker'], strategy)
-        dataset_connectomes.to_csv(output / f"dataset-ds000288_atlas-{atlas_name}_nroi-{nroi}_desc-{strategy_name}_data.tsv", sep='\t')
+        for name in strategy_names:
+            parameters = benchmark_strategies[name]
+            print(f"Denoising: {name}")
+            strategy = {name: parameters}
+            func_data = data_aroma.func if "aroma" in name else data.func
+            dataset_connectomes = deconfound_connectome_single_strategy(func_data, atlas[nroi]['masker'], strategy)
+            dataset_connectomes.to_csv(output / f"dataset-ds000288_atlas-{atlas_name}_nroi-{nroi}_desc-{name}_data.tsv", sep='\t')
 
 
 if __name__ == "__main__":
